fix according to review
